Changed int size to unsigned long for all Benchmarks.
git-svn-id: file:///home/pg512/share/svn/pg512/trunk@648 0f804a50-a62f-0410-84e1-872c020a11f9

+++ benchmark/benchmark.cc
@@ -95,7 +95,7 @@ void Benchmark::evaluate()
     }
 }
 
-void Benchmark::evaluate(int flop)
+void Benchmark::evaluate(unsigned long flop)
 {
     int x = 0;
     double total = 0;
