+++ libla/scaled_sum_BENCHMARK.cc
@@ -18,10 +18,10 @@ class DenseVectorScaledSumBench :
     public Benchmark
 {
     private:
-        int _size;
+        unsigned long _size;
         int _count;
     public:
-        DenseVectorScaledSumBench(const std::string & id, int size, int count) :
+        DenseVectorScaledSumBench(const std::string & id, unsigned long size, int count) :
             Benchmark(id)
         {
             _size = size;
